# Group_84/reports/metrics/openapi.json
{"openapi":"3.1.0","info":{"title":"Loan Approval Risk Service","description":"Real-time consumer-credit risk assessment. Submit an application and receive an approval decision, a calibrated probability and a risk tier.","version":"2.0.0"},"paths":{"/health":{"get":{"tags":["ops"],"summary":"Health","description":"Liveness/readiness probe for Kubernetes.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/v1/model/metadata":{"get":{"tags":["ops"],"summary":"Model Metadata","description":"Expose the served model's identity and its offline training metrics.","operationId":"model_metadata_v1_model_metadata_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Model Metadata V1 Model Metadata Get"}}}}}}},"/v1/predict":{"post":{"tags":["inference"],"summary":"Predict","description":"Score a single loan application.","operationId":"predict_v1_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanApplication"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskAssessmentResponse"}}}},"400":{"description":"Business-rule rejection","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Schema validation failure","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"503":{"description":"Model unavailable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/predict/batch":{"post":{"tags":["inference"],"summary":"Predict Batch","description":"Score up to 100 applications in a single round trip.","operationId":"predict_batch_v1_predict_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchLoanApplications"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchRiskAssessmentResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"503":{"description":"Service Unavailable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"BatchLoanApplications":{"properties":{"applications":{"items":{"$ref":"#/components/schemas/LoanApplication"},"type":"array","maxItems":100,"minItems":1,"title":"Applications"}},"additionalProperties":false,"type":"object","required":["applications"],"title":"BatchLoanApplications","description":"Bounded batch request -- the cap is itself a denial-of-service control."},"BatchRiskAssessmentResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/RiskAssessmentResponse"},"type":"array","title":"Results"},"count":{"type":"integer","title":"Count"},"approved_count":{"type":"integer","title":"Approved Count"}},"type":"object","required":["results","count","approved_count"],"title":"BatchRiskAssessmentResponse","description":"Batch scoring result plus a small roll-up for the caller's dashboard."},"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail"},"error_type":{"type":"string","title":"Error Type"}},"type":"object","required":["detail","error_type"],"title":"ErrorResponse","description":"Uniform error envelope for every non-2xx response."},"HealthResponse":{"properties":{"status":{"type":"string","enum":["healthy","degraded"],"title":"Status"},"model_loaded":{"type":"boolean","title":"Model Loaded"},"model_version":{"type":"string","title":"Model Version"},"environment":{"type":"string","title":"Environment"},"api_version":{"type":"string","title":"Api Version"}},"type":"object","required":["status","model_loaded","model_version","environment","api_version"],"title":"HealthResponse","description":"Liveness/readiness payload consumed by the orchestrator."},"LoanApplication":{"properties":{"age":{"type":"integer","maximum":100.0,"minimum":18.0,"title":"Age","description":"Applicant age in years"},"annual_income":{"type":"integer","maximum":10000000.0,"exclusiveMinimum":0.0,"title":"Annual Income","description":"Gross annual income"},"credit_score":{"type":"integer","maximum":850.0,"minimum":300.0,"title":"Credit Score","description":"Bureau score (300-850)"},"employment_status":{"type":"integer","maximum":2.0,"minimum":0.0,"title":"Employment Status","description":"0=Employed, 1=Self-Employed, 2=Unemployed"},"education_level":{"type":"integer","maximum":4.0,"minimum":0.0,"title":"Education Level","description":"0=Associate, 1=Bachelor, 2=Doctorate, 3=High School, 4=Master"},"loan_amount":{"type":"integer","maximum":10000000.0,"exclusiveMinimum":0.0,"title":"Loan Amount","description":"Principal requested"},"loan_duration":{"type":"integer","maximum":120.0,"minimum":12.0,"title":"Loan Duration","description":"Term in months"},"monthly_debt_payments":{"type":"integer","maximum":1000000.0,"minimum":0.0,"title":"Monthly Debt Payments"},"credit_card_utilization_rate":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Credit Card Utilization Rate"},"debt_to_income_ratio":{"type":"number","maximum":10.0,"minimum":0.0,"title":"Debt To Income Ratio"},"bankruptcy_history":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"Bankruptcy History","description":"0=No, 1=Yes"},"loan_purpose":{"type":"integer","maximum":4.0,"minimum":0.0,"title":"Loan Purpose","description":"0=Auto, 1=Debt Consolidation, 2=Education, 3=Home, 4=Other"},"previous_loan_defaults":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"Previous Loan Defaults","description":"0=No, 1=Yes"},"payment_history":{"type":"integer","maximum":600.0,"minimum":0.0,"title":"Payment History","description":"Months of clean history"},"length_of_credit_history":{"type":"integer","maximum":90.0,"minimum":0.0,"title":"Length Of Credit History"},"savings_account_balance":{"type":"integer","maximum":100000000.0,"minimum":0.0,"title":"Savings Account Balance"},"checking_account_balance":{"type":"integer","maximum":100000000.0,"minimum":0.0,"title":"Checking Account Balance"},"total_liabilities":{"type":"integer","maximum":100000000.0,"minimum":0.0,"title":"Total Liabilities"},"job_tenure":{"type":"integer","maximum":70.0,"minimum":0.0,"title":"Job Tenure"},"net_worth":{"type":"integer","title":"Net Worth","description":"Total assets minus total liabilities"}},"additionalProperties":false,"type":"object","required":["age","annual_income","credit_score","employment_status","education_level","loan_amount","loan_duration","monthly_debt_payments","credit_card_utilization_rate","debt_to_income_ratio","bankruptcy_history","loan_purpose","previous_loan_defaults","payment_history","length_of_credit_history","savings_account_balance","checking_account_balance","total_liabilities","job_tenure","net_worth"],"title":"LoanApplication","description":"One consumer-credit application submitted for scoring.","example":{"age":45,"annual_income":95000,"bankruptcy_history":0,"checking_account_balance":8000,"credit_card_utilization_rate":0.25,"credit_score":720,"debt_to_income_ratio":0.15,"education_level":4,"employment_status":0,"job_tenure":10,"length_of_credit_history":15,"loan_amount":25000,"loan_duration":36,"loan_purpose":3,"monthly_debt_payments":400,"net_worth":220000,"payment_history":28,"previous_loan_defaults":0,"savings_account_balance":35000,"total_liabilities":30000}},"RiskAssessmentResponse":{"properties":{"is_approved":{"type":"boolean","title":"Is Approved","description":"Decision at the configured threshold"},"probability":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Probability","description":"P(approved)"},"risk_tier":{"type":"string","enum":["LOW","MEDIUM","HIGH"],"title":"Risk Tier"},"net_worth":{"type":"integer","title":"Net Worth"},"debt_to_income":{"type":"number","title":"Debt To Income"},"latency_ms":{"type":"number","title":"Latency Ms","description":"Server-side scoring latency"},"model_version":{"type":"string","title":"Model Version"}},"type":"object","required":["is_approved","probability","risk_tier","net_worth","debt_to_income","latency_ms","model_version"],"title":"RiskAssessmentResponse","description":"Scoring result returned to the caller."}}}}